refactor : 카드 조회수 로직 수정

+++ boot/api/src/main/kotlin/com/clip/api/card/controller/dto/CardDetailResponse.kt
@@ -39,7 +39,7 @@ class FeedCardDetailResponse(
     val tags: List<TagResponse>,
     val isOwnCard: Boolean,
     val isFeedCard: Boolean,
-    val visitedCnt: Int,
+    val visitedCnt: Long,
 ) : CardDetailResponse(
     cardId = cardId,
     likeCnt = likeCnt,
@@ -74,7 +74,7 @@ class CommentCardDetailResponse(
     val previousCardId: String,
     val isPreviousCardDeleted: Boolean,
     val previousCardImgUrl: String?,
-    val visitedCnt: Int,
+    val visitedCnt: Long,
 ) : CardDetailResponse(
     cardId = cardId,
     likeCnt = likeCnt,
